Allow only super admins to purge tables on multisites

+++ classes/class-uninstall.php
@@ -71,7 +71,7 @@ class Uninstall {
 		 * AND being uninstalled from the main blog. Otherwise only delete
 		 * data relative to the current blog.
 		 */
-		if ( ! is_multisite() || ( $this->plugin->is_network_activated() && is_main_site() ) ) {
+		if ( ! is_multisite() || ( $this->plugin->is_network_activated() && is_main_site() && is_super_admin() ) ) {
 			$this->delete_all_records();
 			$this->delete_all_options();
 			$this->delete_all_user_meta();
